requirements.txt
streamlit==1.37.0
pandas==1.5.3
scikit-learn==1.5.1
mysql-connector-python==9.3.0
Pillow==10.4.0
numpy>=1.24.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
email-validator 
requests 
streamlit-option-menu==0.3.2
streamlit-float
st-pages 